refactor ixo specifics to own package
