+++ src/Cache.ts
@@ -1,5 +1,5 @@
 var Memcached = require('memcached');
-var dateFormat = require('dateformat');
+import { dateTimeLogger } from './ixo/common/shared';
 var cache: any;
 
 export class Cache {
@@ -8,10 +8,6 @@ export class Cache {
     constructor() {
         this.host = (process.env.MEMCACHE_URI || '');
         cache = new Memcached(this.host, {reconnect: 5000, timeout: 1000, retry: 5000, retries: 1});
-    }
-
-    dateTimeLogger(): string {
-        return dateFormat(new Date(), "yyyy-mm-dd hh:mm:ss:l");
     }
 
     connect(): any {
@@ -36,7 +32,7 @@ export class Cache {
         var inst: any;
         inst = this;
         cache.set(key, value, ttl, function (err: any) {
-            if (err) console.log(inst.dateTimeLogger() + ' Memcache could not set value for key ' + key);
+            if (err) console.log(dateTimeLogger() + ' Memcache could not set value for key ' + key);
         });
     }
