+++ src/handlers/AbstractHandler.ts
@@ -18,18 +18,13 @@ import { SovrinUtils } from '../crypto/SovrinUtils';
 import mq from '../MessageQ';
 import { IWalletModel } from "../model/project/Wallet";
 import { AxiosResponse } from "axios";
-
 import Cache from '../Cache';
 
-var dateFormat = require('dateformat');
+import { dateTimeLogger } from '../ixo/common/shared';
 
 var wallet: IWalletModel;
 
 export abstract class AbstractHandler {
-
-  public dateTimeLogger(): string {
-    return dateFormat(new Date(), "yyyy-mm-dd hh:mm:ss:l");
-  }
 
   public createTransaction(args: any, method: string, model: Model<any>, verifyData?: Function, projectDid?: string) {
 
@@ -56,31 +51,31 @@ export abstract class AbstractHandler {
           })
           return capabilityMap;
         }).catch((reason) => {
-          console.log(this.dateTimeLogger() + ' capabilities not found for project' + request.projectDid);
+          console.log(dateTimeLogger() + ' capabilities not found for project' + request.projectDid);
           reject(new TransactionError('Capabilities not found for project'));
         })
         .then((capabilityMap: any) => {
-          console.log(this.dateTimeLogger() + ' have capability ' + capabilityMap.capability);
+          console.log(dateTimeLogger() + ' have capability ' + capabilityMap.capability);
           TemplateUtils.getTemplateFromCache(capabilityMap.template, request.template)
             .then((schema: any) => {
-              console.log(this.dateTimeLogger() + ' validate the template');
+              console.log(dateTimeLogger() + ' validate the template');
               var validator: ValidatorResult;
               validator = validateJson(schema, args);
               if (validator.valid) {
-                console.log(this.dateTimeLogger() + ' validate the capability');
+                console.log(dateTimeLogger() + ' validate the capability');
                 var capValid: RequestValidator;
                 capValid = request.verifyCapability(capabilityMap.allow);
                 if (capValid.valid) {
-                  console.log(this.dateTimeLogger() + ' verify the signature');
+                  console.log(dateTimeLogger() + ' verify the signature');
                   request.verifySignature(this.preVerifyDidSignature.bind(this), capabilityMap.validateKYC, capabilityMap.capability)
                     .then((sigValid: RequestValidator) => {
                       if (sigValid.valid) {
-                        console.log(this.dateTimeLogger() + ' signature verified');
+                        console.log(dateTimeLogger() + ' signature verified');
                         if (mq.connection != null) {
                           if (verifyData) {
                             verifyData(request)
                               .then(() => {
-                                console.log(this.dateTimeLogger() + ' write transaction to log')
+                                console.log(dateTimeLogger() + ' write transaction to log')
                                 transactionService.createTransaction(request.body, request.signature.type,
                                   request.signature.signatureValue, request.projectDid, capabilityMap.capability)
                                   .then((transaction: ITransactionModel) => {
@@ -91,25 +86,25 @@ export abstract class AbstractHandler {
                                       _created: request.signature.created,
                                       version: request.version + 1
                                     };
-                                    console.log(this.dateTimeLogger() + ' updating the capabilities');
-                                    this.updateCapabilities(request, capabilityMap.capability);
-                                    console.log(this.dateTimeLogger() + ' commit to Elysian');
+                                    console.log(dateTimeLogger() + ' updating the capabilities');
+                                    this.updateCapabilities(request);
+                                    console.log(dateTimeLogger() + ' commit to Elysian');
                                     resolve(model.create({ ...obj, projectDid: request.projectDid }));
-                                    console.log(this.dateTimeLogger() + ' publish to blockchain');
-                                    this.msgToPublish(obj, request, capabilityMap.capability)
+                                    console.log(dateTimeLogger() + ' publish to blockchain');
+                                    this.msgToPublish(obj, request)
                                       .then((msg: any) => {
-                                        console.log(this.dateTimeLogger() + ' message to be published ' + msg.msgType);
+                                        console.log(dateTimeLogger() + ' message to be published ' + msg.msgType);
                                         mq.publish(msg);
                                       });
                                     model.emit('postCommit', obj, request.projectDid);
-                                    console.log(this.dateTimeLogger() + ' transaction completed successfully');
+                                    console.log(dateTimeLogger() + ' transaction completed successfully');
                                   });
                               })
                               .catch((error: string) => {
                                 reject(new TransactionError(error));
                               })
                           } else {
-                            console.log(this.dateTimeLogger() + ' write transaction to log');
+                            console.log(dateTimeLogger() + ' write transaction to log');
                             transactionService.createTransaction(request.body, request.signature.type, request.signature.signatureValue,
                               request.projectDid, capabilityMap.capability)
                               .then((transaction: ITransactionModel) => {
@@ -119,22 +114,22 @@ export abstract class AbstractHandler {
                                   _creator: request.signature.creator,
                                   _created: request.signature.created
                                 };
-                                console.log(this.dateTimeLogger() + ' updating the capabilities');
-                                inst.updateCapabilities(request, capabilityMap.capability);
-                                console.log(this.dateTimeLogger() + ' commit to Elysian');
+                                console.log(dateTimeLogger() + ' updating the capabilities');
+                                inst.updateCapabilities(request);
+                                console.log(dateTimeLogger() + ' commit to Elysian');
                                 resolve(model.create({ ...obj, projectDid: request.projectDid }));
-                                console.log(this.dateTimeLogger() + ' publish to blockchain');
-                                this.msgToPublish(obj, request, capabilityMap.capability)
+                                console.log(dateTimeLogger() + ' publish to blockchain');
+                                this.msgToPublish(obj, request)
                                   .then((msg: any) => {
-                                    console.log(this.dateTimeLogger() + ' message to be published ' + msg.msgType);
+                                    console.log(dateTimeLogger() + ' message to be published ' + msg.msgType);
                                     mq.publish(msg);
                                   });
                                 model.emit('postCommit', obj, request.projectDid);
-                                console.log(this.dateTimeLogger() + ' transaction completed successfully');
+                                console.log(dateTimeLogger() + ' transaction completed successfully');
                               });
                           }
                         } else {
-                          console.log(this.dateTimeLogger() + ' mq currently unavailable');
+                          console.log(dateTimeLogger() + ' mq currently unavailable');
                           reject(new TransactionError('mq currently unavailable'));
                         }
                       } else {
@@ -149,7 +144,7 @@ export abstract class AbstractHandler {
               };
             })
             .catch((reason) => {
-              console.log(this.dateTimeLogger() + 'template registry failed' + reason);
+              console.log(dateTimeLogger() + 'template registry failed' + reason);
               reject(new TransactionError('Cannot connect to template registry'));
             });
         });
@@ -174,31 +169,31 @@ export abstract class AbstractHandler {
           })
           return capabilityMap;
         }).catch((reason) => {
-          console.log(this.dateTimeLogger() + 'capabilities not found for project' + request.projectDid);
+          console.log(dateTimeLogger() + 'capabilities not found for project' + request.projectDid);
           reject(new TransactionError('Capabilities not found for project'));
         })
         .then((capabilityMap: any) => {
-          console.log(this.dateTimeLogger() + ' have capability ' + capabilityMap.capability);
+          console.log(dateTimeLogger() + ' have capability ' + capabilityMap.capability);
           TemplateUtils.getTemplateFromCache(capabilityMap.template, request.template)
             .then((schema: any) => {
-              console.log(this.dateTimeLogger() + ' validate the template');
+              console.log(dateTimeLogger() + ' validate the template');
               var validator: ValidatorResult;
               validator = validateJson(schema, args);
               if (validator.valid) {
-                console.log(this.dateTimeLogger() + ' validate the capability');
+                console.log(dateTimeLogger() + ' validate the capability');
                 var capValid: RequestValidator;
                 capValid = request.verifyCapability(capabilityMap.allow);
                 if (capValid.valid) {
-                  console.log(this.dateTimeLogger() + ' verify the signature');
+                  console.log(dateTimeLogger() + ' verify the signature');
                   request.verifySignature(this.preVerifyDidSignature.bind(this), capabilityMap.validateKYC, capabilityMap.capability)
                     .then((sigValid: RequestValidator) => {
                       if (sigValid.valid) {
-                        console.log(this.dateTimeLogger() + ' query Elysian');
+                        console.log(dateTimeLogger() + ' query Elysian');
                         resolve(query(request.data));
                       } else {
                         reject(new ValidationError(sigValid.errors[0]));
                       }
-                      console.log(this.dateTimeLogger() + ' transaction completed successfully');
+                      console.log(dateTimeLogger() + ' transaction completed successfully');
                     })
                 } else {
                   reject(new ValidationError(capValid.errors[0]));
@@ -208,7 +203,7 @@ export abstract class AbstractHandler {
               };
             })
             .catch((reason) => {
-              console.log(this.dateTimeLogger() + 'template registry failed' + reason);
+              console.log(dateTimeLogger() + 'template registry failed' + reason);
               reject(new TransactionError('Cannot connect to template registry'));
             });
         });
@@ -216,7 +211,7 @@ export abstract class AbstractHandler {
   }
 
 
-  preVerifyDidSignature(didResponse: AxiosResponse, data: Request, capability: string): boolean {
+  preVerifyDidSignature = (didResponse: AxiosResponse, data: Request, capability: string): boolean => {
     return true;
   }
 
@@ -241,15 +236,15 @@ export abstract class AbstractHandler {
         .then((resp: IWalletModel) => {
           wallet = resp;
           Cache.set(wallet.did, { publicKey: wallet.verifyKey });
-          console.log(this.dateTimeLogger() + ' project wallet created');
+          console.log(dateTimeLogger() + ' project wallet created');
           resolve(wallet.did);
         });
     });
   }
 
-  abstract updateCapabilities(request: Request, methodCall: string): void;
+  abstract updateCapabilities = (request: Request): void => {};
 
-  abstract msgToPublish(obj: any, request: Request, methodCall: string): any;
+  abstract msgToPublish = (obj: any, request: Request): any => {};
 
   getWallet(): IWalletModel {
     return wallet;
@@ -290,7 +285,7 @@ export abstract class AbstractHandler {
 
   async publishMessageToQueue(message: any) {
     return new Promise((resolve: Function, reject: Function) => {
-      console.log(this.dateTimeLogger() + ' message to be published ' + message.msgType);
+      console.log(dateTimeLogger() + ' message to be published ' + message.msgType);
       resolve(mq.publish(message));
     });
   }
