+++ src/ixo/model/AgentStatusModel.ts
@@ -1,0 +1,11 @@
+import { Document, Schema, Model, model } from "mongoose";
+
+export interface IAgentStatusModel extends Document { }
+
+var AgentStatusSchema: Schema = new Schema({}, { strict: false });
+
+AgentStatusSchema.pre("save", function (next) {
+  next();
+});
+
+export const AgentStatus: Model<IAgentStatusModel> = model<IAgentStatusModel>("AgentStatus", AgentStatusSchema);
