+++ src/ixo/model/ClaimModel.ts
@@ -1,0 +1,11 @@
+import { Document, Schema, Model, model } from "mongoose";
+
+export interface IClaimModel extends Document { }
+
+var ClaimSchema: Schema = new Schema({}, { strict: false });
+
+ClaimSchema.pre("save", function (next) {
+  next();
+});
+
+export const Claim: Model<IClaimModel> = model<IClaimModel>("Claim", ClaimSchema);
