+++ src/ixo/model/EvaluateClaimModel.ts
@@ -1,0 +1,11 @@
+import { Document, Schema, Model, model } from "mongoose";
+
+export interface IEvaluateClaimModel extends Document { }
+
+var EvaluateClaimSchema: Schema = new Schema({}, { strict: false });
+
+EvaluateClaimSchema.pre("save", function (next) {
+  next();
+});
+
+export const EvaluateClaim: Model<IEvaluateClaimModel> = model<IEvaluateClaimModel>("EvaluateClaim", EvaluateClaimSchema);
