+++ src/routes/QueryRouter.ts
@@ -1,4 +1,3 @@
-import {Router} from 'express';
 import {AbstractRouter} from './AbstractRouter';
 import {QueryHandler} from '../handlers/QueryHandler'
